- Further supports SDCCT-26, SDCCT-33, and SDCCT-34. - Refactored persistence layer to separate entity IDs / resource logical IDs, automatically assign/increment resource instance properties, and (initially) canonicalize resource content. - Refactored resource (parameter) metadata processing + utilization. - Switched to using the sdc-schema-package GitHub repository (2.0-M1 tag) for RFD/SDC XML schemas + RFD WSDL files. - Implemented custom Spring XML namespaces for internal project configuration.

+++ sdcct-core/src/main/java/gov/hhs/onc/sdcct/rfd/RfdResourceRepository.java
@@ -2,9 +2,8 @@ package gov.hhs.onc.sdcct.rfd;
 
 import gov.hhs.onc.sdcct.data.db.SdcctRepository;
 import gov.hhs.onc.sdcct.rfd.impl.RfdResourceImpl;
-import java.math.BigInteger;
 import org.springframework.data.repository.RepositoryDefinition;
 
-@RepositoryDefinition(domainClass = RfdResourceImpl.class, idClass = BigInteger.class)
+@RepositoryDefinition(domainClass = RfdResourceImpl.class, idClass = Long.class)
 public interface RfdResourceRepository extends SdcctRepository<RfdResource> {
 }
